requirements.txt
dj-database-url==0.3.0
Django==1.9.1
gunicorn==19.4.5
psycopg2==2.6.1
SQLAlchemy==1.0.11
whitenoise==2.0.6
markdown==2.6.5

-e git://github.com/botify-labs/django-medusa.git@cc6b1e6#egg=django-medusa
celery==3.1.20
django-pipeline==1.6.4
boto==2.39.0
